requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
colorama==0.4.6
prompt_toolkit==3.0.43
rich==13.7.0
setuptools==69.0.3
pyinstaller==6.2.0 